+++ CHANGELOG.md
@@ -254,6 +254,8 @@
 - Ingest RAEng Newton fund data from IATI
 
 ## [unreleased]
+- Reports can be submitted
+- Submitted reports are shown to users
 
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-14...HEAD
 [release-14]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-13...release-14